bashed it around such that it is reasonably sensible
not entirely sure about using d {
   name: function() {}
}

as a strategy. Means everything is an anonymous function and re-created each time, seems a bit memory hungry

I was rather pleased that it was Limbo like in it's format before I realised.

Javascripts this. semantics have never sat easy with me.

Anyway, we'll see.